Fix bf16 data encoding in FMHA benchmark

+++ projects/composablekernel/tile_engine/ops/fmha/fmha_benchmark.py
@@ -184,7 +184,7 @@ def main():
 
     dtype_map = {
         "fp16": np.float16,
-        "bf16": np.float16,
+        "bf16": np.float32,
         "fp32": np.float32,
         "fp8bf16": np.float16,
         "fp8fp32": np.float16,
